requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.3
Django==4.2.10
django-bootstrap-v5==1.0.11
djangorestframework==3.14.0
mysqlclient==2.2.4
pytz==2024.1
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
